add energy span theory analysis for #4, and move the network drawer to analysis directory
